Updated int loop index to long
This was made to support longer iterations to remove noise and measure data better.

+++ avx_performance.c
@@ -9,11 +9,11 @@
 #include <sys/time.h>
 #include <stdio.h>
 
-void calculate(void (*add)(void*, int, double), int iterations, int type, void*buf) {
+void calculate(void (*add)(void*, int, double), long iterations, int type, void*buf) {
   struct timeval start;
   struct timeval end;
   gettimeofday(&start, NULL);
-  for(int i = 0; i < iterations; i++) {
+  for(long i = 0; i < iterations; i++) {
     double addend = (start.tv_usec / 3600) % (i + 10);
     (*add)(buf, i % (32 / type), addend);
   }
@@ -50,29 +50,29 @@ int main(int argc, char **argv) {
   }
 
   int type;
-  int iterations;
+  long iterations;
   sscanf(argv[1],"%d",&type);
-  sscanf(argv[2],"%d",&iterations);
+  sscanf(argv[2],"%lu",&iterations);
 
   switch(type) {
     case 1:       // operations based on char
-      printf("You've chosen char type and %i iterations.\n", iterations);
+      printf("You've chosen char type and %lu iterations.\n", iterations);
       char charBuf[32] = {3, 1, 7, 5, 2, 4, 6, 1, 3, 0, 6, 6, 8, 2, 1, 2,
                          4, 6, 1, 2, 8, 5, 2, 3, 5, 4, 5, 7, 8, 6, 5, 3};
       calculate(&addChar, iterations, type, &charBuf);
       break;
     case 2:       // operations based on int
-      printf("You've chosen int type and %i iterations.\n", iterations);
+      printf("You've chosen int type and %lu iterations.\n", iterations);
       int intBuf[16] = {3, 1, 7, 5, 2, 4, 6, 1, 3, 0, 6, 6, 8, 2, 1, 2};
       calculate(&addInt, iterations, type, &intBuf);
       break;
     case 4:       //operations based on long
-      printf("You've chosen long type and %i iterations.\n", iterations);
+      printf("You've chosen long type and %lu iterations.\n", iterations);
       long longBuf[8] = {3, 1, 7, 5, 2, 4, 6, 1};
       calculate(&addLong, iterations, type, &longBuf);
       break;
     case 8:       // operations based on double
-      printf("You've chosen double type and %i iterations.\n", iterations);
+      printf("You've chosen double type and %lu iterations.\n", iterations);
       double doubleBuf[4] = {3, 1, 7, 5};
       calculate(&addDouble, iterations, type, &doubleBuf);
       break;
